Removed celery task handler
The celery and redis task handlers were over-engineered for the volumes of traffic we get on the site
and what we need them to do. The original intent was to return from a http request quickly before making
any further requests so I have just used the native python threading module to achieve the same result for now.