Add a healthcheck to postgres and use it in the web service
The Django startup script needs access to the databaseto perform
any migrations. By default docker-compose depends_on only
waits for a service to be up and the service may not be
ready for connections. The web service now waits for postgres
to be healthy as defined by the new healthcheck.