Overhaul run_web_tests to run with existing env/DB
The old script stomped on .env files and did no cleanup.
The new script uses and existing .env/DB if it exists and
generates one a .env file if not. If a .env file and DB did
not exist then they are cleaned at the end.

Services are also only booted if they are not already running.