Remove the mantid cpp stack trace filtering
After testing, I decided the stack trace filtering provided by handel_compressed_cpp_traces.py creates a risk that
some relevant threads are missed out when sending the message to Slack or posting on Github.
Since this was more of a nice-to-have feature I'm removing it for now. If it becomes clear that the full pystack output
is too much for developers to look through, than we can come back to this.