Support cpp stack traces (#60)
* Add new model field cppCompressedTraces

cppCompressedTRaces will be stoed as a base64 compressed string to avoid complications I was having with sending compressed bytes to django.
Had to up the max length on mantidVersion, my dev build with '+uncommited' was reaching 36 characters.

* imporve trimmed down formatting

* move cpp trace code to utils module and add to github issue maker

* flake8 changes

* fix name being used as email

* add stack trace as attachment text so it collapses

* add check to avoid out of bounds error

* remove unused logger

* add precaution for core files without file paths

* move model constants to file

* fix bug in handle_compressed_cpp_traces and add some tests

* fix bug in ghub issue manager

* Remove the mantid cpp stack trace filtering

After testing, I decided the stack trace filtering provided by handel_compressed_cpp_traces.py creates a risk that
some relevant threads are missed out when sending the message to Slack or posting on Github.
Since this was more of a nice-to-have feature I'm removing it for now. If it becomes clear that the full pystack output
is too much for developers to look through, than we can come back to this.

* import from services not web